# docs/swagger.yaml
basePath: /api/v1
definitions:
  handler.ArticleStatusRequest:
    properties:
      is_later:
        example: false
        type: boolean
      is_read:
        example: true
        type: boolean
    type: object
  model.Article:
    properties:
      content:
        description: 記事本文（省略可能）
        type: string
      created_at:
        description: 作成日時
        type: string
      feed_id:
        description: 所属フィードID
        type: string
      id:
        description: 記事の一意識別子
        type: string
      is_later:
        description: 後で見るフラグ
        type: boolean
      is_read:
        description: 既読フラグ
        type: boolean
      published_at:
        description: 公開日時
        type: string
      title:
        description: 記事タイトル
        type: string
      url:
        description: 記事の元URL
        type: string
    type: object
  model.Feed:
    properties:
      created_at:
        description: 作成日時
        type: string
      folder_id:
        description: 所属フォルダID
        type: string
      id:
        description: フィードの一意識別子
        type: string
      last_updated:
        description: 最終更新日時
        type: string
      name:
        description: フィード名（必須）
        type: string
      plugin_type:
        description: プラグイン種別（必須）
        type: string
      update_interval:
        description: 更新間隔（分）
        type: integer
      url:
        description: フィードURL（必須、URL形式）
        type: string
    required:
    - name
    - plugin_type
    - url
    type: object
  model.Folder:
    properties:
      created_at:
        description: 作成日時
        type: string
      id:
        description: フォルダの一意識別子
        type: string
      name:
        description: フォルダ名（必須）
        type: string
      user_id:
        description: 将来のマルチユーザー対応用
        type: string
    required:
    - name
    type: object
host: localhost:8080
info:
  contact:
    email: support@feedapp.example.com
    name: FeedApp Support
    url: http://www.feedapp.example.com/support
  description: RSS フィードリーダーとカスタムスクレイピング機能を提供するAPIサーバー
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: FeedApp API
  version: "1.0"
paths:
  /articles:
    get:
      consumes:
      - application/json
      description: データベースに保存されているすべての記事を取得します
      produces:
      - application/json
      responses:
        "200":
          description: 記事一覧
          schema:
            items:
              $ref: '#/definitions/model.Article'
            type: array
        "500":
          description: サーバー内部エラー
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 記事一覧取得
      tags:
      - articles
  /articles/{id}:
    get:
      consumes:
      - application/json
      description: 指定されたIDの記事を取得します
      parameters:
      - description: 記事ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 記事詳細
          schema:
            $ref: '#/definitions/model.Article'
        "404":
          description: 記事が見つかりません
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: サーバー内部エラー
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 記事詳細取得
      tags:
      - articles
  /articles/{id}/status:
    put:
      consumes:
      - application/json
      description: 指定されたIDの記事の読了状態や後で読む状態を更新します
      parameters:
      - description: 記事ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: 記事状態情報
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/handler.ArticleStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新された記事
          schema:
            $ref: '#/definitions/model.Article'
        "400":
          description: リクエストボディの形式が不正
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 記事が見つかりません
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: サーバー内部エラー
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 記事状態更新
      tags:
      - articles
  /articles/later:
    get:
      consumes:
      - application/json
      description: 「後で読む」に設定された記事の一覧を取得します
      produces:
      - application/json
      responses:
        "200":
          description: 後で読む記事一覧
          schema:
            items:
              $ref: '#/definitions/model.Article'
            type: array
        "500":
          description: サーバー内部エラー
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 後で読む記事一覧取得
      tags:
      - articles
  /feeds:
    get:
      consumes:
      - application/json
      description: データベースに保存されているすべてのフィードを取得します
      produces:
      - application/json
      responses:
        "200":
          description: フィード一覧
          schema:
            items:
              $ref: '#/definitions/model.Feed'
            type: array
        "500":
          description: サーバー内部エラー
          schema:
            additionalProperties:
              type: string
            type: object
      summary: フィード一覧取得
      tags:
      - feeds
    post:
      consumes:
      - application/json
      description: 新しいフィードを作成します
      parameters:
      - description: フィード情報
        in: body
        name: feed
        required: true
        schema:
          $ref: '#/definitions/model.Feed'
      produces:
      - application/json
      responses:
        "201":
          description: 作成されたフィード
          schema:
            $ref: '#/definitions/model.Feed'
        "400":
          description: リクエストボディの形式が不正
          schema:
            additionalProperties: true
            type: object
        "500":
          description: サーバー内部エラー
          schema:
            additionalProperties:
              type: string
            type: object
      summary: フィード作成
      tags:
      - feeds
  /feeds/{id}:
    delete:
      consumes:
      - application/json
      description: 指定されたIDのフィードを削除します
      parameters:
      - description: フィードID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: 削除成功
        "404":
          description: フィードが見つかりません
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: サーバー内部エラー
          schema:
            additionalProperties:
              type: string
            type: object
      summary: フィード削除
      tags:
      - feeds
    get:
      consumes:
      - application/json
      description: 指定されたIDのフィードを取得します
      parameters:
      - description: フィードID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: フィード詳細
          schema:
            $ref: '#/definitions/model.Feed'
        "404":
          description: フィードが見つかりません
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: サーバー内部エラー
          schema:
            additionalProperties:
              type: string
            type: object
      summary: フィード詳細取得
      tags:
      - feeds
    put:
      consumes:
      - application/json
      description: 指定されたIDのフィードを更新します
      parameters:
      - description: フィードID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: 更新するフィード情報
        in: body
        name: feed
        required: true
        schema:
          $ref: '#/definitions/model.Feed'
      produces:
      - application/json
      responses:
        "200":
          description: 更新されたフィード
          schema:
            $ref: '#/definitions/model.Feed'
        "400":
          description: リクエストボディの形式が不正
          schema:
            additionalProperties: true
            type: object
        "404":
          description: フィードが見つかりません
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: サーバー内部エラー
          schema:
            additionalProperties:
              type: string
            type: object
      summary: フィード更新
      tags:
      - feeds
  /folders:
    get:
      consumes:
      - application/json
      description: データベースに保存されているすべてのフォルダを取得します
      produces:
      - application/json
      responses:
        "200":
          description: フォルダ一覧
          schema:
            items:
              $ref: '#/definitions/model.Folder'
            type: array
        "500":
          description: サーバー内部エラー
          schema:
            additionalProperties:
              type: string
            type: object
      summary: フォルダ一覧取得
      tags:
      - folders
    post:
      consumes:
      - application/json
      description: 新しいフォルダを作成します
      parameters:
      - description: フォルダ情報
        in: body
        name: folder
        required: true
        schema:
          $ref: '#/definitions/model.Folder'
      produces:
      - application/json
      responses:
        "201":
          description: 作成されたフォルダ
          schema:
            $ref: '#/definitions/model.Folder'
        "400":
          description: リクエストボディの形式が不正
          schema:
            additionalProperties: true
            type: object
        "500":
          description: サーバー内部エラー
          schema:
            additionalProperties:
              type: string
            type: object
      summary: フォルダ作成
      tags:
      - folders
  /folders/{id}:
    delete:
      consumes:
      - application/json
      description: 指定されたIDのフォルダを削除します
      parameters:
      - description: フォルダID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: 削除成功
        "404":
          description: フォルダが見つかりません
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: サーバー内部エラー
          schema:
            additionalProperties:
              type: string
            type: object
      summary: フォルダ削除
      tags:
      - folders
    get:
      consumes:
      - application/json
      description: 指定されたIDのフォルダを取得します
      parameters:
      - description: フォルダID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: フォルダ詳細
          schema:
            $ref: '#/definitions/model.Folder'
        "404":
          description: フォルダが見つかりません
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: サーバー内部エラー
          schema:
            additionalProperties:
              type: string
            type: object
      summary: フォルダ詳細取得
      tags:
      - folders
    put:
      consumes:
      - application/json
      description: 指定されたIDのフォルダを更新します
      parameters:
      - description: フォルダID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: 更新するフォルダ情報
        in: body
        name: folder
        required: true
        schema:
          $ref: '#/definitions/model.Folder'
      produces:
      - application/json
      responses:
        "200":
          description: 更新されたフォルダ
          schema:
            $ref: '#/definitions/model.Folder'
        "400":
          description: リクエストボディの形式が不正
          schema:
            additionalProperties: true
            type: object
        "404":
          description: フォルダが見つかりません
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: サーバー内部エラー
          schema:
            additionalProperties:
              type: string
            type: object
      summary: フォルダ更新
      tags:
      - folders
schemes:
- http
- https
swagger: "2.0"
tags:
- description: フォルダ管理API
  name: folders
- description: フィード管理API
  name: feeds
- description: 記事管理API
  name: articles